Refactor the spaced repetition items to use the FSRS CSharp implementation instead of the python one. Removing the python dependency completly

+++ Avalonia/Avalonia.Flecs/Avalonia.Flecs.StellaLearning/Converters/SpacedRepetitionItemConverter.cs
@@ -145,7 +145,7 @@ namespace Avalonia.Flecs.StellaLearning.Converters
                             baseItem.Priority = reader.GetInt32();
                             break;
                         case SpacedRepetitionItem baseItem when propertyName == nameof(SpacedRepetitionItem.Step):
-                            baseItem.Step = reader.TokenType == JsonTokenType.Null ? null : reader.GetInt64();
+                            baseItem.Step = reader.TokenType == JsonTokenType.Null ? null : (int)reader.GetInt64();
                             break;
                         case SpacedRepetitionItem baseItem when propertyName == nameof(SpacedRepetitionItem.LastReview):
                             baseItem.LastReview = reader.TokenType == JsonTokenType.Null ? null : reader.GetDateTime();
